[RSPEC TESTS] for Regulation Form APIs: moved general stuff in context
